refactor: move breakpoints to theme
