+++ src/components/PricingCard/PricingCard.js
@@ -15,7 +15,8 @@ import colors from "../../theme/colors"
 import fonts from "../../theme/fonts"
 import fontSizes from "../../theme/fontSizes"
 import fontWeights from "../../theme/fontWeights"
-import { breakpoints, radius } from "../../utils/presets"
+import { radius } from "../../utils/presets"
+import breakpoints from "../../theme/breakpoints"
 import space from "../../theme/space"
 import { Heading } from "../Heading"
