+++ src/components/SettingsBlock/SettingsBlock.js
@@ -7,7 +7,8 @@ import { MdHelpOutline } from "react-icons/md"
 import { ContentBox } from "../ContentBox"
 import { LinkButton } from "../LinkButton"
 import { Heading } from "../Heading"
-import { breakpoints, fontFamilies } from "../../utils/presets"
+import { fontFamilies } from "../../utils/presets"
+import breakpoints from "../../theme/breakpoints"
 import space from "../../theme/space"
 import cardStyles from "../../theme/styles/card"
 import fontSizes from "../../theme/fontSizes"
