+++ src/components/Tabs/TabsNav.js
@@ -2,7 +2,8 @@
 import { jsx } from "@emotion/core"
 import PropTypes from "prop-types"
 
-import { breakpoints, dimensions } from "../../utils/presets"
+import { dimensions } from "../../utils/presets"
+import breakpoints from "../../theme/breakpoints"
 
 const TabsNav = ({ children }) => (
   <nav
