+++ src/index.ts
@@ -83,9 +83,9 @@ export { default as colors } from "./theme/colors"
 export { default as space } from "./theme/space"
 export { default as fontSizes } from "./theme/fontSizes"
 export { default as fontWeights } from "./theme/fontWeights"
+export { default as breakpoints } from "./theme/breakpoints"
 
 export { fontFamilies } from "./utils/presets/typography"
-export { breakpoints } from "./utils/presets/breakpoints"
 export { dimensions, radius } from "./utils/presets/sizes"
 export { zIndices } from "./utils/presets/zIndices"
