+++ src/theme/breakpoints.d.ts
@@ -1,0 +1,5 @@
+type Threshold = `mobile` | `phablet` | `tablet` | `desktop` | `hd`
+
+const breakpoints: Record<Threshold, number>
+
+export = breakpoints
