+++ src/theme/breakpoints.js
@@ -1,7 +1,9 @@
-export const breakpoints = {
+const breakpoints = {
   mobile: 360,
   phablet: 550,
   tablet: 750,
   desktop: 1000,
   hd: 1300,
 }
+
+export default breakpoints
