+++ src/utils/presets/index.ts
@@ -1,5 +1,4 @@
 export { fontFamilies } from "./typography"
-export { breakpoints } from "./breakpoints"
 export { dimensions, radius } from "./sizes"
 export { zIndices } from "./zIndices"
 export { transition } from "./transition"
