+++ src/utils/presets/typography.ts
@@ -38,20 +38,3 @@ export const fontFamilies = {
     `monospace`,
   ].join(),
 }
-
-export const fontSizes = {
-  "7xl": `4.5rem`,
-  "6xl": `3.75rem`,
-  "5xl": `3rem`,
-  "4xl": `2.625rem`,
-  "3xl": `2rem`,
-  "2xl": `1.75rem`,
-  xl: `1.5rem`,
-  l: `1.25rem`,
-  m: `1.125rem`,
-  s: `1rem`,
-  xs: `0.875rem`,
-  "2xs": `0.75rem`,
-  "3xs": `0.6875rem`,
-  "4xs": `0.5rem`,
-}
